Code restructuring for more modular design and testing. Added test for conversion from dataframe to inventory
